Fixed errors

+++ users/views.py
@@ -2,42 +2,44 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.auth.views import LoginView
 from django.contrib import auth, messages
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.core.cache import cache
 from django.db.models import Prefetch
 from django.http import HttpResponseRedirect
 from django.shortcuts import redirect
 from django.urls import reverse, reverse_lazy
-from django.views.generic import CreateView, UpdateView, TemplateView
+from django.views.generic import CreateView, TemplateView, UpdateView
 from carts.models import Cart
-
 from common.mixins import CacheMixin
 from orders.models import Order, OrderItem
 
 from users.forms import ProfileForm, UserLoginForm, UserRegistrationForm
 
 
-def update_cart_for_user(session_key, user):
-    if session_key:
-        Cart.objects.filter(session_key=session_key).update(user=user)
-        forgot_carts = Cart.objects.filter(user=user)
-        if forgot_carts.exists():
-            forgot_carts.delete()
-
-
 class UserLoginView(LoginView):
     template_name = "users/login.html"
     form_class = UserLoginForm
+    # success_url = reverse_lazy('main:index')
 
     def get_success_url(self):
-        return self.request.POST.get("next") or reverse_lazy("main:index")
+        redirect_page = self.request.POST.get("next", None)
+        if redirect_page and redirect_page != reverse("user:logout"):
+            return redirect_page
+        return reverse_lazy("main:index")
 
     def form_valid(self, form):
         session_key = self.request.session.session_key
-        user = form.get_user()
-        auth.login(self.request, user)
-        update_cart_for_user(session_key, user)
-        messages.success(self.request, f"{user.username}, Вы вошли в аккаунт")
-        return HttpResponseRedirect(self.get_success_url())
+        if user := form.get_user():
+            auth.login(self.request, user)
+            if session_key:
+                # delete old authorized user carts
+                forgot_carts = Cart.objects.filter(user=user)
+                if forgot_carts.exists():
+                    forgot_carts.delete()
+                # add new authorized user carts from anonimous session
+                Cart.objects.filter(session_key=session_key).update(user=user)
+
+                messages.success(self.request, f"{user.username}, Вы вошли в аккаунт")
+
+                return HttpResponseRedirect(self.get_success_url())
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -52,14 +54,20 @@ class UserRegistrationView(CreateView):
 
     def form_valid(self, form):
         session_key = self.request.session.session_key
-        user = form.save()
-        auth.login(self.request, user)
-        update_cart_for_user(session_key, user)
+        user = form.instance
+
+        if user:
+            form.save()
+            auth.login(self.request, user)
+
+        if session_key:
+            Cart.objects.filter(session_key=session_key).update(user=user)
+
         messages.success(
             self.request,
             f"{user.username}, Вы успешно зарегистрированы и вошли в аккаунт",
         )
-        return HttpResponseRedirect(self.get_success_url())
+        return HttpResponseRedirect(self.success_url)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -87,6 +95,7 @@ class UserProfileView(LoginRequiredMixin, CacheMixin, UpdateView):
         context = super().get_context_data(**kwargs)
         context["title"] = "Home - Кабинет"
 
+        # Можно вынести сам запрос в отдельный метод этого класса контроллера
         orders = (
             Order.objects.filter(user=self.request.user)
             .prefetch_related(
@@ -99,7 +108,7 @@ class UserProfileView(LoginRequiredMixin, CacheMixin, UpdateView):
         )
 
         context["orders"] = self.set_get_cache(
-            orders, f"orders_{self.request.user.pk}", 60 * 2
+            orders, f"user_{self.request.user.id}_orders", 60
         )
         return context
 
@@ -115,6 +124,6 @@ class UserCartView(TemplateView):
 
 @login_required
 def logout(request):
-    messages.success(request, f"{request.user.username} Вы вышли из аккаунта")
+    messages.success(request, f"{request.user.username}, Вы вышли из аккаунта")
     auth.logout(request)
     return redirect(reverse("main:index"))
